extract feed parsing and NewsList from NewsContainer
