--- source/views/news/parse-feed.js
@@ -1,6 +1,0 @@
-// @flow
-import {parseString} from 'xml2js'
-import pify from 'pify'
-import type {FeedResponseType} from './types'
-
-export const parseXml: ((text: string) => FeedResponseType) = pify(parseString)
